feat: add compat for legacy message
