+++ packages/i18n/src/locales/ja.json
@@ -2421,7 +2421,10 @@
             "Prompt": {
                 "SchemaContextTitle": "[現在のデータコンテキスト]",
                 "DatabaseContextTitle": "[データベースコンテキスト]",
-                "CopilotContextTitle": "[現在の Copilot コンテキスト]"
+                "CopilotContextTitle": "[現在の Copilot コンテキスト]",
+                "ExecutionIntent": {
+                    "Policy": "現在のターンの実行ポリシー:\n- sqlRunner を使うかどうかを決める前に、最新のユーザーメッセージの意図を分類してください。\n- ユーザーが SQL の生成、表示、作成のみを明示している場合、またはクエリを実行しないよう求めている場合は、このターンでは sqlRunner を呼び出さず、確認が必要な場合を除いて SQL だけを返してください。\n- ユーザーが表示、一覧、ランキング、比較、集計、グラフ作成、または実際のデータベース結果の取得を求めており、sqlRunner が利用可能な場合は、最終回答の前に読み取り専用 SQL で sqlRunner を呼び出してください。\n- データ結果が必要なリクエストでは、SQL を書くだけで終了しないでください。\n- 探索的または明細の結果件数は 100 行以下にしてください。"
+                }
             },
             "ChartBuilder": {
                 "Description": "提供されたデータに基づいてグラフ設定を返します。\n\n使い方:\n1. chartType（bar/line/area/pie）を指定します。\n2. data はクエリ結果行の配列です。\n3. xKey / yKeys / categoryKey / valueKey が不明な場合は空にしてください。ツールが推定します。\n   - Line/area: x には時間フィールド、y には数値フィールドを優先します。\n   - Bar: x にはカテゴリフィールド、y には数値フィールドを優先します。\n   - Pie: categoryKey にカテゴリフィールド、valueKey に数値フィールドを選択します。\n4. ツールは時間/数値/カテゴリ列を推定し、必要に応じて options.xKeyType/sortBy を設定します。",
